fixes from review
