+++ src/main/java/org/gridsuite/study/server/dto/CreatedStudyBasicInfos.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.study.server.dto.elasticsearch;
+package org.gridsuite.study.server.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
