+++ src/main/java/org/gridsuite/study/server/dto/StudyInfos.java
@@ -4,34 +4,21 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.study.server.dto.elasticsearch;
+package org.gridsuite.study.server.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 import lombok.experimental.SuperBuilder;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.elasticsearch.annotations.Field;
-
-import java.util.UUID;
 
 /**
- * @author Slimane Amar <slimane.amar at rte-france.com>
+ * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
  */
 @SuperBuilder
 @NoArgsConstructor
 @Getter
-@ToString
-@EqualsAndHashCode
-@Schema(description = "Basic study attributes")
-public class BasicStudyInfos {
-    @Id
-    String uniqueId;
-
-    @Field("studyUuid")
-    UUID id;
-
-    String userId;
+@ToString(callSuper = true)
+@Schema(description = "Study attributes")
+public class StudyInfos extends CreatedStudyBasicInfos {
 }
