+++ src/main/java/org/gridsuite/study/server/elasticsearch/StudyInfosService.java
@@ -8,7 +8,7 @@ package org.gridsuite.study.server.elasticsearch;
 
 import co.elastic.clients.elasticsearch._types.query_dsl.QueryStringQuery;
 
-import org.gridsuite.study.server.dto.elasticsearch.CreatedStudyBasicInfos;
+import org.gridsuite.study.server.dto.CreatedStudyBasicInfos;
 import org.springframework.data.elasticsearch.client.elc.NativeQuery;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.data.elasticsearch.core.SearchHit;
