+++ src/test/java/org/gridsuite/study/server/StudyServiceTest.java
@@ -14,7 +14,7 @@ import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.exceptions.UncheckedInterruptedException;
 import com.powsybl.network.store.client.NetworkStoreService;
 
-import org.gridsuite.study.server.dto.elasticsearch.BasicStudyInfos;
+import org.gridsuite.study.server.dto.BasicStudyInfos;
 import org.gridsuite.study.server.notification.NotificationService;
 import org.gridsuite.study.server.repository.StudyRepository;
 import org.gridsuite.study.server.service.CaseService;
