+++ src/test/java/org/gridsuite/study/server/StudyTest.java
@@ -33,10 +33,7 @@ import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 import okio.Buffer;
 import org.gridsuite.study.server.dto.*;
-import org.gridsuite.study.server.dto.elasticsearch.BasicStudyInfos;
-import org.gridsuite.study.server.dto.elasticsearch.CreatedStudyBasicInfos;
 import org.gridsuite.study.server.dto.elasticsearch.EquipmentInfos;
-import org.gridsuite.study.server.dto.elasticsearch.StudyInfos;
 import org.gridsuite.study.server.dto.modification.ModificationInfos;
 import org.gridsuite.study.server.dto.modification.ModificationType;
 import org.gridsuite.study.server.elasticsearch.EquipmentInfosService;
