+++ src/test/java/org/gridsuite/study/server/utils/MatcherStudyInfos.java
@@ -8,7 +8,7 @@ package org.gridsuite.study.server.utils;
 
 import java.util.UUID;
 
-import org.gridsuite.study.server.dto.elasticsearch.StudyInfos;
+import org.gridsuite.study.server.dto.StudyInfos;
 
 /**
  * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
